make sure that we don't push non-linear history to github
once we switch to open source version of the repo we don't want old
commits to accidentally get pushed to the public repo. The best way to
ensude that would be to check if branch is an ancestor of the initial
commit 767..., but because of the way gitlab works I can't figure out
the way to do that. So instead we are going to check if the branch is an
ancestor of the new main branch we created for our open source history
(abf48a5e0) and only sync the branches that are.